Move the interactive tools to a subpackage
The command line and full screen interfaces, the session they drive and the
parsing they need are all there to debug a device rather than to use the
library, so they move together into pyaprilaire.cli, next to the mock server.
The library itself is now just the client, the socket, the packets and the
constants.

The entry point moves to __main__ as it did for the mock server, so
`python -m pyaprilaire.cli` and the `pyaprilaire` command are unchanged, as
are the module's arguments and behavior.

frame moves too. It reads the wire protocol, which looks like library
material, but it exists to show frames that Packet.parse discards and
nothing outside the tools uses it.
