+++ pyaprilaire/cli/__main__.py
@@ -1,0 +1,10 @@
+"""Entry point for running an interactive session"""
+
+from __future__ import annotations
+
+import sys
+
+from . import main
+
+if __name__ == "__main__":
+    sys.exit(main())
